Add emitSig() to send signals back to client
Adds emitSig() to LMOS so that signals and events can be sent back via
the parser to the client as strings.

bp_waitForReply only returned the ID number if there was a matching
ID in the list. bp_waitForString returns the message string instead,
which is more useful.

Adds getMessageAt and deleteMessageAt functions to allow random
access to the message list.

This is required to fix the issue whereby polling for messages
removed them from the list, resulting in the loss of any
out-of-order messages.

Should resolve #7, but more testing required.