Add Auto Generated Id
Adds AutoGenerateId(), which returns an unused id. This is used
by bp_sendCommandAndWait() if an id is not supplied. A struct is now
used to return the exit code, along with the generated id if used.

This fixes the issue of having to create ids manually, which is not
needed most of the time, if ever.

Resolves #10.