Create Rough IDS Capture Client
Adds a simple client to capture from the IDS camera and save to PNG. The
implementation needs more work but can at least capture an image.

The compiler has been switched to zig cc with the musl c library, since
it allows us to create binaries that can be directly copied to the
camera virtual machine without glibc linking issues.