Convert Inkscape Plugin to New Message System
Begins changing laser.py to use the new Blastpit functions for queueing
and uploading commands. Since laser.py is a real mess, and will be
hugely simplified when moving over to the geometry engine, this is just
a quick fix for now.

Also adds functions for queueing messages for a single upload.