Implement Multiple Message Parsing
Gets the multiple message parsing to a state ready for fuzzing tests.

The Inkscape plugin can now send drawings and qpset commands. It is now
possible to wait for a specific reply or timeout.

A mutex was added to Parser because of an issue caused by the fact that
sending a message automatically recieves messages, triggering the
sending of more replies followed by more receipts. Eventually this
causes the stack to fill up. Now we queue replies and use the mutex to
prevent the timer triggering while we deal with the current messages.