Fix test failure due to missing shell command
Checks for the existence of the 'passed' shell command instead of
assuming that it is available.

The 'passed' command uses figlet to print the word 'passed' in bold
letters upon success, but is not available outside of our local build
system. This change simply tests to see if the command is available
before running it.